Move FFTDimension into its own module
